+++ packages/docs/qui-docs/types.d.ts
@@ -2,16 +2,6 @@
 /// <reference types="@qualcomm-ui/mdx-vite/docs-plugin" />
 /// <reference types="@qualcomm-ui/mdx-vite/react-demo-plugin" />
 
-declare module "*.mdx" {
-  let MDXComponent: (props: any) => React.ReactNode
-  export const frontmatter: {
-    category?: string
-    id: string
-    title: string
-  }
-  export default MDXComponent
-}
-
 declare module "@qualcomm-ui/mdx-vite-plugin" {
   import type {SiteData} from "@qualcomm-ui/mdx-common"