+++ packages/frameworks/react-swagger/src/internal/auth/authorize-operation-btn.tsx
@@ -1,0 +1,45 @@
+// Modified from https://github.com/swagger-api/swagger-ui
+// Apache-2.0
+// Changes from Qualcomm Technologies, Inc. are provided under the following license:
+// Copyright (c) Qualcomm Technologies, Inc. and/or its subsidiaries.
+// SPDX-License-Identifier: BSD-3-Clause-Clear
+
+import type {MouseEvent, ReactNode} from "react"
+
+import {LockIcon, LockOpenIcon} from "lucide-react"
+
+import {IconButton} from "@qualcomm-ui/react/button"
+
+AuthorizeOperationBtn.displayName = "AuthorizeOperationBtn"
+
+export interface AuthorizeOperationBtnProps {
+  getComponent: () => any
+  isAuthorized: boolean
+  onClick?: () => void
+}
+
+export function AuthorizeOperationBtn(
+  props: AuthorizeOperationBtnProps,
+): ReactNode {
+  function onClick(e: MouseEvent) {
+    e.stopPropagation()
+
+    props.onClick?.()
+  }
+
+  const {isAuthorized} = props
+
+  return (
+    <IconButton
+      aria-label={
+        isAuthorized
+          ? "authorization button locked"
+          : "authorization button unlocked"
+      }
+      icon={isAuthorized ? LockIcon : LockOpenIcon}
+      onClick={onClick}
+      size="sm"
+      variant="ghost"
+    />
+  )
+}