+++ packages/frameworks/react-swagger/src/internal/json-model-viewer/components/data-types/boolean.tsx
@@ -1,0 +1,21 @@
+// Copyright (c) Qualcomm Technologies, Inc. and/or its subsidiaries.
+// SPDX-License-Identifier: BSD-3-Clause-Clear
+
+import {defineEasyType} from "./define-easy-type"
+
+export const booleanType = defineEasyType<boolean>({
+  colorKey: "base0E",
+  deserialize: (value) => {
+    if (value === "true") {
+      return true
+    }
+    if (value === "false") {
+      return false
+    }
+    throw new Error("Invalid boolean value")
+  },
+  is: (value) => typeof value === "boolean",
+  Renderer: ({value}) => <>{value ? "true" : "false"}</>,
+  serialize: (value) => value.toString(),
+  type: "bool",
+})